Sprint 9. HttpServer. CodeStyle. To review
