--- src/httpserver/LocalDurationTypeAdapter.java
@@ -1,21 +1,0 @@
-package httpserver;
-
-import com.google.gson.TypeAdapter;
-import com.google.gson.stream.JsonReader;
-import com.google.gson.stream.JsonWriter;
-
-import java.io.IOException;
-import java.time.Duration;
-
-class LocalDurationTypeAdapter extends TypeAdapter<Duration> {
-
-    @Override
-    public void write(JsonWriter jsonWriter, Duration duration) throws IOException {
-        jsonWriter.value(duration.toMinutes());
-    }
-
-    @Override
-    public Duration read(JsonReader jsonReader) throws IOException {
-        return Duration.ofMinutes(jsonReader.nextLong());
-    }
-}
